Write test files and re-organize the project's directory structure
