+++ lib/server_pages.rb
