+++ lib/views/academic.erb
