+++ lib/views/info.erb
